refactor!: Give the extras types their natural namespaces
The types shipped alongside Uno.UI for both Skia and WinAppSDK carried the
assembly name as their namespace. They now sit where they would have been
written had they lived in Uno.UI directly, mirroring the WinUI namespace of
whatever each one extends:

  ElevatedView, CommandBarExtensions,
  SplitViewExtensions            -> Uno.UI.Xaml.Controls
  UIElementExtensions            -> Uno.UI.Xaml
  AutomationPropertiesExtensions -> Uno.UI.Xaml.Automation
  VisibleBoundsPadding           -> Uno.UI.Behaviors
  StorageFileHelper              -> Uno.Storage
  RectExtensions (internal)      -> Uno.UI.Extensions
  DevTools.Input/Xaml (internal) -> Uno.UI.DevTools.Input/Xaml

VisibleBoundsPadding now declares one namespace for both legs, so the IS_UNO
fork only switches the class name. The padding and reflection helpers move out
of UIElementExtensions into Uno.UI.Extensions.UIElementExtensions, gated to the
WinAppSDK build so the Uno legs use the Uno.UI copy instead of a duplicate.
