+++ src/Uno.UI.Extras/CommandBarExtensions.cs
@@ -5,7 +5,7 @@ using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Media;
 
-namespace Uno.UI.Extras
+namespace Uno.UI.Xaml.Controls
 {
 	public static class CommandBarExtensions
 	{
