+++ src/Uno.UI.Extras/DevTools/Input/Finger.cs
@@ -16,7 +16,7 @@ using System.Runtime.InteropServices;
 using PointerDeviceType = Windows.Devices.Input.PointerDeviceType;
 #endif
 
-namespace Uno.UI.Extras.DevTools.Input;
+namespace Uno.UI.DevTools.Input;
 
 internal partial class Finger : IInjectedPointer, IDisposable
 {
