+++ src/Uno.UI.Extras/DevTools/Input/InputInjectorExtensions.cs
@@ -14,7 +14,7 @@ using System.Runtime.InteropServices;
 using PointerDeviceType = Windows.Devices.Input.PointerDeviceType;
 #endif
 
-namespace Uno.UI.Extras.DevTools.Input;
+namespace Uno.UI.DevTools.Input;
 
 internal static class InputInjectorExtensions
 {
