+++ src/Uno.UI.Extras/DevTools/Input/PointExtensions.cs
@@ -13,7 +13,7 @@ using System.Runtime.InteropServices;
 using PointerDeviceType = Windows.Devices.Input.PointerDeviceType;
 #endif
 
-namespace Uno.UI.Extras.DevTools.Input;
+namespace Uno.UI.DevTools.Input;
 
 internal static class PointExtensions
 {
