+++ src/Uno.UI.Extras/LinkerDefinition.net6.0.xml
@@ -1,8 +1,8 @@
-﻿<linker>
+<linker>
   <assembly fullname="Uno.UI.Extras">
-	<type fullname="Uno.UI.Extras.CommandBarExtensions"/>
-	<type fullname="Uno.UI.Extras.UIElementExtensions"/>
-	<type fullname="Uno.UI.Extras.SplitViewExtensions"/>
+	<type fullname="Uno.UI.Xaml.Controls.CommandBarExtensions"/>
+	<type fullname="Uno.UI.Xaml.UIElementExtensions"/>
+	<type fullname="Uno.UI.Xaml.Controls.SplitViewExtensions"/>
 	<type fullname="Uno.UI.Markup.FromJsonExtension"/>
   </assembly>
 </linker>
