+++ src/Uno.UI.Extras/SplitViewExtensions.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 
-namespace Uno.UI.Extras
+namespace Uno.UI.Xaml.Controls
 {
 	public static class SplitViewExtensions
 	{
