+++ src/Uno.UI.Extras/Themes/Generic.xaml
@@ -1,12 +1,12 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 					xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-					xmlns:extras="using:Uno.UI.Extras">
+					xmlns:controls="using:Uno.UI.Xaml.Controls">
 
-	<Style TargetType="extras:ElevatedView">
+	<Style TargetType="controls:ElevatedView">
 		<Setter Property="IsTabStop" Value="False" />
 		<Setter Property="Template">
 			<Setter.Value>
-				<ControlTemplate TargetType="extras:ElevatedView">
+				<ControlTemplate TargetType="controls:ElevatedView">
 					<Grid>
 						<Canvas x:Name="PART_ShadowHost" />
