+++ src/Uno.UI/Helpers/XamlHelper.cs
@@ -26,7 +26,9 @@ internal static partial class XamlHelper
 	{
 		[string.Empty] = "http://schemas.microsoft.com/winfx/2006/xaml/presentation",
 		["x"] = "http://schemas.microsoft.com/winfx/2006/xaml",
-		["extras"] = "using:Uno.UI.Extras", // uno utilities
+		["uno"] = "using:Uno.UI.Xaml.Controls", // uno controls and control extensions
+		["unoxaml"] = "using:Uno.UI.Xaml", // uno UIElement extensions
+		["behaviors"] = "using:Uno.UI.Behaviors", // uno behaviors
 		["muxc"] = "using:Microsoft.UI.Xaml.Controls",
 	};
